auth-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.2.1
email-validator==2.1.1
